refactor(callbacks): extract Overview core helpers to _callbacks_overview.py
Step 7a of the callbacks.py decomposition (issue #87). First Overview
sub-PR; two more will follow once this lands (panels block + briefing
block) to keep diffs reviewable.

## Re-scoping note

Original decomposition plan estimated Overview at ~600 lines. Closer
inspection puts it at ~1,900 lines across 25 helpers — too big for a
single PR. Splitting into 3 sub-PRs:

* **7a (this PR):** hero / metrics / insight block — 5 functions, ~300
  lines.
* **7b:** panels (drivers / generation / models leaderboard / risk /
  scenarios) — ~600 lines.
* **7c:** briefing surface (sparklines / briefing / digest / spotlights
  / weather context / data-health / changes / news / persona KPIs) —
  ~900 lines.

After all three land the file owns every ``_build_overview_*`` /
``_spotlight_*`` / ``_build_persona_kpis`` helper.

## What moves in this PR

Five functions from ``components/callbacks.py`` into
``components/_callbacks_overview.py``:

* ``_build_overview_title`` — page-title block (region name + subtitle).
* ``_build_overview_metrics_items`` — 5-up MetricsBar (Now / 7d Peak /
  7d Low / Average / 24h Trend), NaN-aware so the hero metric never
  displays as ``nan`` when EIA-930 has a publishing lag.
* ``_build_overview_hero_chart`` — 7d actual demand + 24h forecast
  bridge with 80% confidence band (mirrors v2 DemandChart.tsx).
* ``_build_overview_model_card`` — horizontal model-performance bar
  (MAPE / RMSE / MAE / R² + trained/simulated badge).
* ``_build_overview_insight`` — three-sentence narrative with semantic-
  color delta spans (rising demand reads as "warning" not "good").

Also dropped a now-stale breadcrumb comment in callbacks.py that
pointed to the extracted ``_latest_real_demand`` reference.

## Public-import surface preserved

``components/callbacks.py`` re-imports each function by name.
``register_callbacks`` continues to call them directly through the
same callbacks-module namespace. Tests using ``from components.callbacks
import _build_overview_metrics_items`` continue to resolve.

Added a ``# noqa: F401`` marker on the ``_latest_real_demand`` re-export
from shared — once ``_build_overview_insight`` moved out, it's no
longer called by callbacks.py itself, only re-exported for test
imports.

## Test patches

No patches updated — none of the Overview tests patch internal
collaborators; they call the helpers directly.

## Size impact

callbacks.py: 5,668 → 5,369 (–299 lines)
_callbacks_overview.py: new, 383 lines

Cumulative progress on the decomposition:
* PR #98 (shared):       7,167 → 7,051 (–116)
* PR #99 (us_grid):      7,051 → 6,389 (–662)
* PR #100 (models):      6,389 → 6,183 (–206)
* PR #101 (alerts):      6,183 → 5,978 (–205)
* PR #102 (generation):  5,978 → 5,802 (–176)
* PR #103 (weather):     5,802 → 5,668 (–134)
* this PR (overview 7a): 5,668 → 5,369 (–299)

Total: −1,798 lines from the monolith across 7 PRs (−25.1%).

## Tests

All 1,568 unit tests pass locally. Ruff clean.
